more cleanup, more tests
