+++ .github/workflows/python-app.yml
@@ -37,4 +37,4 @@ jobs:
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Test with pytest
       run: |
-        python -m unittest discover .
+        QT_QPA_PLATFORM=offscreen PYDM_DEFAULT_PROTOCOL=fake python -m unittest discover .
