+++ backend/runner.py
@@ -2,25 +2,20 @@ from datetime import datetime
 from time import sleep
 
 from backend.backend_cavity import BackendCavity
-from backend.backend_cryomodule import BackendCryomodule
-from backend.backend_ssa import BackendSSA
-
 from lcls_tools.common.controls.pyepics.utils import PV
-from lcls_tools.superconducting.sc_linac import Machine
+from lcls_tools.superconducting.sc_linac import Machine, Cryomodule
 from lcls_tools.superconducting.sc_linac_utils import ALL_CRYOMODULES
 from utils.utils import DEBUG, BACKEND_SLEEP_TIME
 
 WATCHER_PV: PV = PV("PHYS:SYS0:1:SC_CAV_FAULT_HEARTBEAT")
 WATCHER_PV.put(0)
 
-DISPLAY_MACHINE = Machine(
-    ssa_class=BackendSSA, cavity_class=BackendCavity, cryomodule_class=BackendCryomodule
-)
+DISPLAY_MACHINE = Machine(cavity_class=BackendCavity)
 
 while True:
     start = datetime.now()
     for cryomoduleName in ALL_CRYOMODULES:
-        cryomodule: BackendCryomodule = DISPLAY_MACHINE.cryomodules[cryomoduleName]
+        cryomodule: Cryomodule = DISPLAY_MACHINE.cryomodules[cryomoduleName]
         for cavity in cryomodule.cavities.values():
             cavity.run_through_faults()
     if DEBUG:
