+++ frontend/cavity_widget.py
@@ -1,16 +1,20 @@
 from dataclasses import dataclass
+from typing import Optional
 
 import numpy as np
-from PyQt5.QtCore import pyqtSignal
-from PyQt5.QtGui import QColor, QCursor, QFont, QFontMetrics, QPainter, QPen
+from PyQt5.QtCore import pyqtSignal, QPoint
+from PyQt5.QtGui import (
+    QColor,
+    QCursor,
+    QFont,
+    QFontMetrics,
+    QPainter,
+    QPen,
+    QMouseEvent,
+)
 from pydm import Display, PyDMChannel
 from pydm.widgets.drawing import PyDMDrawingPolygon
 from qtpy.QtCore import Property as qtProperty, QRect, Qt, Slot
-
-from backend.backend_cryomodule import BackendCryomodule
-from frontend.cavity_fault_display import FaultCavity
-from lcls_tools.common.frontend.display.util import showDisplay
-from lcls_tools.superconducting.sc_linac import Machine
 
 GREEN_FILL_COLOR = QColor(9, 141, 0)
 YELLOW_FILL_COLOR = QColor(244, 230, 67)
@@ -23,8 +27,6 @@ LIMEGREEN_FILL_COLOR = QColor(92, 253, 92)
 BLACK_TEXT_COLOR = QColor(0, 0, 0)
 DARK_GRAY_COLOR = QColor(40, 40, 40)
 WHITE_TEXT_COLOR = QColor(250, 250, 250)
-
-FAULT_MACHINE = Machine(cavity_class=FaultCavity, cryomodule_class=BackendCryomodule)
 
 
 @dataclass
@@ -46,63 +48,48 @@ SHAPE_PARAMETER_DICT = {
 
 
 class CavityWidget(PyDMDrawingPolygon):
-    pressPos = None
+    press_pos: Optional[QPoint] = None
     clicked = pyqtSignal()
 
     def __init__(self, parent=None, init_channel=None):
         super(CavityWidget, self).__init__(parent, init_channel)
         self._num_points = 4
-        self._cavityText = "TEXT"
+        self._cavity_text = ""
         self._underline = False
         self._pen = QPen(BLACK_TEXT_COLOR)  # Shape's border color
         self._rotation = 0
         self._brush.setColor(QColor(201, 255, 203))  # Shape's fill color
-        self._pen.setWidth(1.0)
-        self._severity_channel: PyDMChannel = None
-        self._description_channel: PyDMChannel = None
+        self._pen.setWidth(1)
+        self._severity_channel: Optional[PyDMChannel] = None
+        self._description_channel: Optional[PyDMChannel] = None
         self.alarmSensitiveBorder = False
         self.alarmSensitiveContent = False
         self._faultDisplay: Display = None
-        self.clicked.connect(self.show_fault_display)
-        self.cavityNumber = None
-        self.cmName = None
         self.setCursor(QCursor(Qt.PointingHandCursor))
 
-    def show_fault_display(self):
-        showDisplay(self.faultDisplay)
-
     # The following two functions were copy/pasted from stack overflow
-    def mousePressEvent(self, event):
+    def mousePressEvent(self, event: QMouseEvent):
         if event.button() == Qt.LeftButton:
-            self.pressPos = event.pos()
+            self.press_pos = event.pos()
 
-    def mouseReleaseEvent(self, event):
+    def mouseReleaseEvent(self, event: QMouseEvent):
         # ensure that the left button was pressed *and* released within the
         # geometry of the widget; if so, emit the signal;
         if (
-            self.pressPos is not None
+            self.press_pos is not None
             and event.button() == Qt.LeftButton
             and event.pos() in self.rect()
         ):
             self.clicked.emit()
-        self.pressPos = None
-
-    @property
-    def faultDisplay(self):
-        if not self._faultDisplay:
-            fault_cavity: FaultCavity = FAULT_MACHINE.cryomodules[self.cmName].cavities[
-                self.cavityNumber
-            ]
-            self._faultDisplay = fault_cavity.display
-        return self._faultDisplay
+        self.press_pos = None
 
     @qtProperty(str)
-    def cavityText(self):
-        return self._cavityText
+    def cavity_text(self):
+        return self._cavity_text
 
-    @cavityText.setter
-    def cavityText(self, text):
-        self._cavityText = text
+    @cavity_text.setter
+    def cavity_text(self, text):
+        self._cavity_text = text
 
     @qtProperty(str)
     def description_channel(self):
@@ -137,17 +124,17 @@ class CavityWidget(PyDMDrawingPolygon):
 
     @Slot(int)
     def severity_channel_value_changed(self, value: int):
-        self.changeShape(
+        self.change_shape(
             SHAPE_PARAMETER_DICT[value]
             if value in SHAPE_PARAMETER_DICT
             else SHAPE_PARAMETER_DICT[3]
         )
 
-    def changeShape(self, shapeParameterObject):
-        self.brush.setColor(shapeParameterObject.fillColor)
-        self.penColor = shapeParameterObject.borderColor
-        self.numberOfPoints = shapeParameterObject.numPoints
-        self.rotation = shapeParameterObject.rotation
+    def change_shape(self, shape_parameter_object):
+        self.brush.setColor(shape_parameter_object.fillColor)
+        self.penColor = shape_parameter_object.borderColor
+        self.numberOfPoints = shape_parameter_object.numPoints
+        self.rotation = shape_parameter_object.rotation
         self.update()
 
     @qtProperty(bool)
@@ -160,16 +147,16 @@ class CavityWidget(PyDMDrawingPolygon):
 
     def value_changed(self, new_val):
         super(CavityWidget, self).value_changed(new_val)
-        self.cavityText = new_val
+        self.cavity_text = new_val
         self.update()
 
     def draw_item(self, painter: QPainter):
         super(CavityWidget, self).draw_item(painter)
         x, y, w, h = self.get_bounds()
-        rect = QRect(x, y, w, h)
+        rect = QRect(int(x), int(y), w, h)
         fm = QFontMetrics(painter.font())
-        if self._cavityText:
-            sx = rect.width() / fm.width(self._cavityText)
+        if self._cavity_text:
+            sx = rect.width() / fm.width(self._cavity_text)
             sy = rect.height() / fm.height()
 
             painter.save()
@@ -179,13 +166,13 @@ class CavityWidget(PyDMDrawingPolygon):
 
             # Text color
             pen = QPen(QColor(240, 240, 240))
-            pen.setWidth(5.0)
+            pen.setWidth(5)
 
             font = QFont()
             font.setUnderline(self._underline)
             painter.setFont(font)
 
             painter.setPen(pen)
-            painter.drawText(rect, Qt.AlignCenter, self._cavityText)
+            painter.drawText(rect, Qt.AlignCenter, self._cavity_text)
             painter.setPen(self._pen)
             painter.restore()
