+++ frontend/gui_cavity.py
@@ -1,28 +1,43 @@
 import json
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Optional
 
+from PyQt5.QtCore import Qt
 from PyQt5.QtGui import QColor
-from PyQt5.QtWidgets import QVBoxLayout, QHBoxLayout, QSizePolicy
-from frontend.cavity_widget import CavityWidget
-from pydm.widgets import PyDMByteIndicator
+from PyQt5.QtWidgets import (
+    QVBoxLayout,
+    QHBoxLayout,
+    QSizePolicy,
+    QGroupBox,
+    QScrollArea,
+    QAbstractScrollArea,
+    QLabel,
+)
+from edmbutton import PyDMEDMDisplayButton
+from pydm import Display
+from pydm.widgets import PyDMByteIndicator, PyDMShellCommand, PyDMRelatedDisplayButton
 
-from lcls_tools.superconducting.sc_linac import Cavity
+from backend.backend_cavity import BackendCavity
+from frontend.cavity_widget import CavityWidget
+from frontend.utils import make_header, EnumLabel, PyDMFaultButton
+from lcls_tools.common.frontend.display.util import showDisplay
 
 if TYPE_CHECKING:
     from lcls_tools.superconducting.sc_linac import Rack
 
 
-class GUICavity(Cavity):
+class GUICavity(BackendCavity):
     def __init__(self, cavity_num: int, rack_object: "Rack"):
         super().__init__(cavity_num, rack_object)
+        self._fault_display: Optional[Display] = None
+        self.fault_display_grid_layout = make_header()
+
         self.vert_layout = QVBoxLayout()
         self.cavity_widget = CavityWidget()
         self.cavity_widget.setMinimumSize(10, 10)
         self.cavity_widget.setAccessibleName("cavity_widget")
-        self.cavity_widget.cavityText = str(cavity_num)
+        self.cavity_widget.cavity_text = str(cavity_num)
         self.cavity_widget.setSizePolicy(QSizePolicy.Preferred, QSizePolicy.Preferred)
-        self.cavity_widget.cmName = self.cryomodule.name
-        self.cavity_widget.cavityNumber = cavity_num
+        self.cavity_widget.clicked.connect(self.show_fault_display)
 
         self.hor_layout = QHBoxLayout()
 
@@ -71,3 +86,84 @@ class GUICavity(Cavity):
         ]
 
         self.ssa_bar.rules = json.dumps(rule)
+
+    def populate_fault_display(self):
+        for idx, fault in enumerate(self.faults.values()):
+            code_label = QLabel(fault.tlc)
+            code_label.setSizePolicy(QSizePolicy.Maximum, QSizePolicy.Maximum)
+            code_label.setAlignment(Qt.AlignCenter)
+
+            short_description_label = QLabel(fault.short_description)
+            short_description_label.setSizePolicy(
+                QSizePolicy.Maximum, QSizePolicy.Maximum
+            )
+            short_description_label.setAlignment(Qt.AlignLeft)
+            short_description_label.setWordWrap(True)
+
+            action_label = QLabel(fault.action)
+            action_label.setSizePolicy(
+                QSizePolicy.MinimumExpanding, QSizePolicy.MinimumExpanding
+            )
+            action_label.setAlignment(Qt.AlignLeft)
+            action_label.setWordWrap(True)
+
+            status_label = EnumLabel(fault=fault, codeLabel=code_label)
+            status_label.setSizePolicy(QSizePolicy.Maximum, QSizePolicy.Maximum)
+
+            row_idx = idx + 1
+            self.fault_display_grid_layout.addWidget(code_label, row_idx, 0)
+            self.fault_display_grid_layout.addWidget(
+                short_description_label, row_idx, 1
+            )
+            self.fault_display_grid_layout.addWidget(status_label, row_idx, 2)
+
+            if fault.button_level == "EDM":
+                button = PyDMEDMDisplayButton()
+                button.filenames = [fault.button_command]
+                button.macros = fault.macros + (
+                    "," + fault.button_macro if fault.button_macro else ""
+                )
+
+            elif fault.button_level == "SCRIPT":
+                button = PyDMShellCommand()
+                button.commands = [fault.button_command]
+
+            elif fault.button_level == "PYDM":
+                button = PyDMFaultButton()
+                button.openInNewWindow = True
+                button.filenames = [fault.button_command]
+                button.macros = self.cryomodule.pydm_macros
+
+            else:
+                button = PyDMRelatedDisplayButton()
+                button.setEnabled(False)
+
+            self.fault_display_grid_layout.addWidget(button, row_idx, 3)
+            self.fault_display_grid_layout.addWidget(action_label, row_idx, 4)
+            button.setText(fault.button_text)
+            button.showIcon = False
+
+    def show_fault_display(self):
+        showDisplay(self.fault_display)
+
+    @property
+    def fault_display(self):
+        if not self._fault_display:
+            groupbox = QGroupBox()
+            groupbox.setLayout(self.fault_display_grid_layout)
+            self.populate_fault_display()
+
+            self._fault_display = Display()
+            self._fault_display.setWindowTitle(f"{self} Faults")
+            vlayout = QVBoxLayout()
+
+            self._fault_display.setLayout(vlayout)
+
+            scroll_area = QScrollArea()
+            scroll_area.setVerticalScrollBarPolicy(Qt.ScrollBarAlwaysOn)
+            scroll_area.setSizeAdjustPolicy(QAbstractScrollArea.AdjustToContents)
+            scroll_area.setWidgetResizable(True)
+            vlayout.addWidget(scroll_area)
+            scroll_area.setWidget(groupbox)
+
+        return self._fault_display
