+++ utils/utils.py
@@ -3,7 +3,7 @@ from csv import DictReader
 
 from typing import Dict, List
 
-DEBUG = True
+DEBUG = False
 BACKEND_SLEEP_TIME = 10
 
 STATUS_SUFFIX = "CUDSTATUS"
